Code Challenges:
Add an album to the ALBUMSDATA array in the JS section. Hit the Run button, then click the Albums link and see how it's changed. This is a demo of how dynamic routing can work.
Next, for the real challenge: You are invited to play around with the code and see what you can add - up to you. Some ideas: Add a component that displays albums by year, using a router parameter/match param for the year. Add a Random link that displays a random album. Display the album title along with "Album 1" and "Album 2". Add images and links to lyrics/MP3s... Up to you!